refactor global stores
