+++ client/src/components/core/Notifications/index.ts
@@ -1,0 +1,1 @@
+export * from './Notifications';
