+++ client/src/components/core/index.ts
@@ -4,5 +4,6 @@ export * from './Dialog';
 export * from './Drawer';
 export * from './LanguageToggle';
 export * from './Link';
+export * from './Notifications';
 export * from './Spinner';
 export * from './Table';
