+++ client/src/components/layout/Sidebar/UserDropup.tsx
@@ -7,8 +7,8 @@ import { HiUserCircle } from 'react-icons/hi2';
 import { IoIosArrowUp } from 'react-icons/io';
 
 import { LanguageToggle } from '@/components/core';
-import { useAuthStore } from '@/features/auth';
 import { useOnClickOutside } from '@/hooks/useOnClickOutside';
+import { useAuthStore } from '@/stores/auth-store';
 
 export const UserDropup = ({ username }: { username?: string }) => {
   const auth = useAuthStore();
