+++ client/src/features/auth/index.ts
@@ -1,2 +1,1 @@
 export * from './pages/LoginPage';
-export * from './stores/auth.store';
