--- client/src/features/notifications/index.ts
@@ -1,2 +1,0 @@
-export * from './components/Notifications';
-export * from './stores/notifications.store';
