+++ client/src/stores/auth-store.ts
